requirements.txt
# AI Red Teaming Toolkit Requirements
flask==3.0.0
flask-cors==4.0.0
openai==1.53.0
anthropic==0.39.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
jinja2==3.1.2
werkzeug==3.0.1
gunicorn==21.2.0
pandas==2.1.4
numpy==1.24.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
beautifulsoup4==4.12.2
nltk==3.8.1
scikit-learn==1.3.2
asyncio==3.4.3
aiohttp==3.9.1
colorama==0.4.6
tqdm==4.66.1
click==8.1.7